requirements.txt
elasticsearch
faiss-cpu
google-genai
langchain
numpy
polars
python-dotenv
sentence-transformers
streamlit
tqdm
logging
sentence-transformers
